Output results from querying randomuser.me
